fixed types again

+++ aruw-mcb-project/src/aruwsrc/control/chassis/new-chassis/mecanum_wheel.hpp
@@ -42,14 +42,14 @@ private:
     const double WHEEL_RELATIVE_TO_ROLLER_ANGLE = M_PI_4;
     const double AXLE_TO_ROBOT_FRONT = M_PI_2;
     const CMSISMat<2, 2> MAT1 = CMSISMat<2, 2>({0.0f,
-                                                sin(WHEEL_RELATIVE_TO_ROLLER_ANGLE),
+                                                (float)sin(WHEEL_RELATIVE_TO_ROLLER_ANGLE),
                                                 config.diameter / 2,
-                                                cos(WHEEL_RELATIVE_TO_ROLLER_ANGLE)})
+                                                (float)cos(WHEEL_RELATIVE_TO_ROLLER_ANGLE)})
                                     .inverse();
-    const CMSISMat<2, 2> MAT2 = CMSISMat<2, 2>({cos(AXLE_TO_ROBOT_FRONT),
-                                                -sin(AXLE_TO_ROBOT_FRONT),
-                                                sin(AXLE_TO_ROBOT_FRONT),
-                                                cos(AXLE_TO_ROBOT_FRONT)})
+    const CMSISMat<2, 2> MAT2 = CMSISMat<2, 2>({(float)cos(AXLE_TO_ROBOT_FRONT),
+                                                (float)-sin(AXLE_TO_ROBOT_FRONT),
+                                                (float)sin(AXLE_TO_ROBOT_FRONT),
+                                                (float)cos(AXLE_TO_ROBOT_FRONT)})
                                     .inverse();
     /// product of matrices 1 and 2 in equation on Swerve! Notion
     const CMSISMat<2, 2> PRODUCT_MAT = MAT1 * MAT2;
